Refactored Menubar component to make it resusable
